Simplified the database set lookup process (not all sets are switched over). Partially updated magic system tables. Database strings like clan names are kept in CString form (until we get rid of MFC!). Implemented SetCString() equiv of SetKOString() (hacky, but it works for now). Removed unnecessary strings from being loaded (e.g. magic table names/descriptions, item names...). CKnights class now calls InitialiseValue() on instantiation, removed duplicate defaults code. CKnights also now defaults members to 1 (as above). Fixed a bug with C3DMap - m_ppnEvent isn't NULL by default, so it'd crash on cleanup if it didn't load events. Fixed a bug with the version list not setting the filename.

+++ Aujard/Define.h
@@ -25,7 +25,7 @@
 
 struct _ITEM_TABLE
 {
-	int   m_iNum;				// item num
+	long  m_iNum;				// item num
 	BYTE  m_bCountable;			// 개수 개념 아이템
 };
